Restructure the net-module
Create new folder 'net'. Creating types.rs for all structs and enums
from original net.rs
