+++ server/src/net/mod.rs
@@ -14,111 +14,15 @@
 //!
 //! ...
 //!
+pub mod types;
 
 use std::io::{Write, Read};
 // to encode and decode the structs to the given stream
-use bincode::rustc_serialize::{
-    decode_from, encode_into, EncodingError, DecodingError
-};
+use bincode::rustc_serialize::{decode_from, encode_into};
 use bincode::SizeLimit;
-use rustc_serialize::{Encodable, Encoder}; // to encode the Error
-use std::io;
+use net::types::*;
 
 const PROTOCOL_VERSION: u8 = 1;
-
-/// Code numeric value sent as first byte
-#[derive(PartialEq, RustcEncodable, RustcDecodable)]
-#[repr(u8)]
-pub enum PkgType {
-    Greet = 0,
-    Login,
-    Command,
-    Error,
-    Ok,
-    Response,
-    AccDenied,
-    AccGranted,
-}
-
-/// Struct to send the kind of error and error message to the client
-#[derive(RustcEncodable, RustcDecodable, Debug)]
-pub struct ClientErrMsg {
-    code: u16,
-    msg: String
-}
-
-/// Convert the possible Error to a serializable ClientErrMsg struct
-impl From<Error> for ClientErrMsg {
-    fn from(error: Error) -> ClientErrMsg {
-        match error {
-            Error::Io(_) => ClientErrMsg {
-                code: 0,
-                msg: "IO error".into()
-            },
-            Error::UnexpectedPkg(err) => ClientErrMsg {
-                code: 2,
-                msg: err.into()
-            },
-            Error::UnknownCmd(err) => ClientErrMsg {
-                code: 3,
-                msg: err.into()
-            },
-            Error::Encode(_) => ClientErrMsg {
-                code: 4,
-                msg: "encoding error".into()
-            },
-            Error::Decode(_) => ClientErrMsg {
-                code: 5,
-                msg: "decoding error".into()
-            }
-        }
-    }
-}
-
-/// Collection of possible errors while communicating with the client
-#[derive(Debug)]
-pub enum Error {
-    Io(io::Error),
-    UnexpectedPkg(String),
-    UnknownCmd(String),
-    Encode(EncodingError),
-    Decode(DecodingError),
-}
-
-/// Implement the conversion from io::Error to NetworkError
-impl From<io::Error> for Error {
-    fn from(err: io::Error) -> Error {
-        Error::Io(err)
-    }
-}
-
-/// Implement the conversion from EncodingError to NetworkError
-impl  From<EncodingError> for Error {
-    fn from(err: EncodingError) -> Error {
-        Error::Encode(err)
-    }
-}
-
-/// Implement the conversion from DecodingError to NetworkError
-impl From<DecodingError> for Error {
-    fn from(err: DecodingError) -> Error {
-        Error::Decode(err)
-    }
-}
-
-/// This is the first packet being sent by the server after the TCP connection
-/// is established.
-#[derive(RustcEncodable, RustcDecodable)]
-pub struct Greeting {
-    pub protocol_version: u8,   // 1 byte
-    pub message: String,        // n bytes
-}
-
-impl Greeting {
-    pub fn make_greeting(version: u8, msg: String) -> Greeting {
-        Greeting { protocol_version: version, message: msg }
-    }
-}
 
 /// Write a welcome-message to the given server-client-stream
 pub fn do_handshake<W: Write + Read>(stream: &mut W)
@@ -136,14 +40,6 @@ pub fn do_handshake<W: Write + Read>(stream: &mut W)
         Ok(sth) => Ok((sth.username, sth.password)),
         Err(msg) => Err(msg)
     }
-}
-
-/// The client responds with this packet to a `Greeting` packet, finishing the
-/// authentication handshake.
-#[derive(Default, RustcEncodable, RustcDecodable)]
-pub struct Login {
-    pub username: String,
-    pub password: String
 }
 
 /// reads the data from the response to the handshake,
@@ -170,20 +66,6 @@ pub fn send_error_package<W: Write>(mut stream: &mut W, err: ClientErrMsg)
     try!(encode_into(&PkgType::Error, stream, SizeLimit::Bounded(1024)));
     try!(encode_into(&err, &mut stream, SizeLimit::Bounded(1024)));
     Ok(())
-}
-
-/// Sent by the client to the server.
-///
-/// Many commands are executed via query, but there are some "special"
-/// commands that are not sent as query.
-#[derive(RustcEncodable, RustcDecodable, Debug, PartialEq)]
-#[repr(u8)]
-pub enum Command {
-    Quit,
-    Ping,
-    Query(String),
-    // Shutdown,
-    // Statistics,
 }
 
 /// Read the sent bytes, extract the kind of command
@@ -217,12 +99,6 @@ pub fn send_ok_packet<W: Write>(mut stream: &mut W)
     try!(encode_into(&PkgType::Ok, stream, SizeLimit::Bounded(1024)));
     Ok(())
 }
-
-/// Sent by the server to the client.
-pub struct Response;
-
-   // TODO
-
 
 // # Some information for the `net` working group:
 //
